Simplify voting routing: remove list page, keep /vote/:pollId only
- Remove /vote list page, only /vote/:pollId route remains
- Delete VotePage list, rename PollVotePage to VotePage
- Remove "All Polls" back link since no list exists
- Use hardcoded pollId (not display name lookup) in ESPN, update for prod
