+++ src/pages/ESPN.tsx
@@ -9,7 +9,7 @@ import { Button } from '@/components/ui/button';
 const matches = [
   {
     name: 'EASTERN FINALS',
-    pollVoteId: '3f6473dc-7a63-4980-babd-cc2dae862cdc',
+    pollId: '3f6473dc-7a63-4980-babd-cc2dae862cdc',
     teams: [
       {
         name: 'Team Boston',
@@ -169,7 +169,7 @@ const matches = [
   },
   {
     name: 'WESTERN FINALS',
-    pollVoteId: 'e73fcd3c-2ed7-4c7c-8a54-996d10cd3a2f',
+    pollId: 'e73fcd3c-2ed7-4c7c-8a54-996d10cd3a2f',
     teams: [
       {
         name: 'Team Toronto',
@@ -530,7 +530,7 @@ export default function ESPNPage() {
             </div>
             <div className="flex justify-center mt-8">
               <Button asChild>
-                <Link to={`/vote/${match.pollVoteId}`}>
+                <Link to={`/vote/${match.pollId}`}>
                   Vote for {match.name} MVP!
                 </Link>
               </Button>
